+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapClient.lua
@@ -9,12 +9,14 @@ Ext.Require("S7_StatsConfigurator.lua")
 --  ===================================
 
 local function S7_StatsLoader() --  Loads stats-configuration json during StatsLoaded Event.
-    local fileName = S7_ConfigSettings.StatsLoader.FileName --  lists all config files.
-    S7_DebugLog("Loading " .. fileName)
-    table.insert(toConfigure, {[fileName] = Ext.LoadFile(fileName)}) --  Queue files for configuration.
-    S7_StatsConfigurator() --  Pass stringified JSON to StatsConfigurator()
-    toConfigure = {} -- flush list
-    S7_DebugLog("StatsLoading Completed.")
+    if S7_ConfigSettings.StatsLoader.Enable == true then
+        local fileName = S7_ConfigSettings.StatsLoader.FileName --  lists all config files.
+        S7_DebugLog("Loading " .. fileName)
+        table.insert(toConfigure, {[fileName] = Ext.LoadFile(fileName)}) --  Queue files for configuration.
+        S7_StatsConfigurator() --  Pass stringified JSON to StatsConfigurator()
+        toConfigure = {} -- flush list
+        S7_DebugLog("StatsLoading Completed.")
+    end
 end
 
 --  ===============================================